+++ src/main/java/act/util/AnnotatedClassFinder.java
@@ -22,10 +22,7 @@ package act.util;
 
 import act.app.event.SysEventId;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * The annotation is used on a certain method to mark it